third_party/WebKit/LayoutTests/fast/scrolling/empty-layer-anchor.html

<!DOCTYPE html>
<div id="teleport" style="position:absolute; top:10000px; overflow:hidden;"></div>
<div style="position:absolute; top:10000px; height:150vh;">PASS</div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
  test(function() {
    window.location.href = "#teleport";
    assert_equals(document.documentElement.scrollTop, 10000);
  }, "Even empty objects should be scrolled into view");
</script>
